USMNdArray returns float64 array by default

+++ numba_dpex/tests/core/types/USMNdArray/test_array_creation_errors.py
@@ -6,7 +6,7 @@ from numba_dpex.core.types import USMNdArray
 
 def test_init():
     usma = USMNdArray(1, device=None, queue=None)
-    assert usma.dtype.name == "int64"
+    assert usma.dtype.name == "float64"
     assert usma.ndim == 1
     assert usma.layout == "C"
     assert usma.addrspace == 1
@@ -19,7 +19,7 @@ def test_init():
     device = dpctl.SyclDevice().filter_string
 
     usma = USMNdArray(1, device=device, queue=None)
-    assert usma.dtype.name == "int64"
+    assert usma.dtype.name == "float64"
     assert usma.ndim == 1
     assert usma.layout == "C"
     assert usma.addrspace == 1
@@ -39,7 +39,7 @@ def test_init():
 
     queue = dpctl.SyclQueue()
     usma = USMNdArray(1, device=None, queue=queue)
-    assert usma.dtype.name == "int64"
+    assert usma.dtype.name == "float64"
     assert usma.ndim == 1
     assert usma.layout == "C"
     assert usma.addrspace == 1
